passed scan options to factory
